requirements.txt
autopep8==1.6.0
bidict==0.21.4
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
gunicorn==20.1.0
h11==0.13.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==9.0.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-engineio==4.3.1
python-socketio==5.5.2
qrcode==7.3.1
simple-websocket==0.5.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
wsproto==1.1.0
